Remove `DefaultGraph` trait
