+++ src/graphs/mod.rs
@@ -7,9 +7,6 @@ pub mod structs;
 mod base;
 pub use base::*;
 
-mod default;
-pub use default::*;
-
 mod direction;
 pub use direction::*;
