+++ src/graphs/structs/undirected_dense_adjacency_matrix.rs
@@ -14,7 +14,7 @@ use serde::{Deserialize, Serialize};
 use crate::{
     graphs::{
         algorithms::traversal::{DFSEdge, DFSEdges, Traversal},
-        directions, BaseGraph, DefaultGraph, PartialOrdGraph, PathGraph, SubGraph, UndirectedGraph,
+        directions, BaseGraph, PartialOrdGraph, PathGraph, SubGraph, UndirectedGraph,
     },
     io::DOT,
     prelude::BFS,
@@ -165,6 +165,17 @@ impl Hash for UndirectedDenseAdjacencyMatrixGraph {
     }
 }
 
+impl Default for UndirectedDenseAdjacencyMatrixGraph {
+    #[inline]
+    fn default() -> Self {
+        Self {
+            labels: Default::default(),
+            adjacency_matrix: DenseAdjacencyMatrix::from_elem((0, 0), false),
+            size: 0,
+        }
+    }
+}
+
 impl BaseGraph for UndirectedDenseAdjacencyMatrixGraph {
     type Data = DenseAdjacencyMatrix;
 
@@ -219,6 +230,69 @@ impl BaseGraph for UndirectedDenseAdjacencyMatrixGraph {
                 size += 1;
             }
         }
+
+        Self {
+            labels,
+            adjacency_matrix,
+            size,
+        }
+    }
+
+    fn empty<V, I>(labels: I) -> Self
+    where
+        V: Into<String>,
+        I: IntoIterator<Item = V>,
+    {
+        // Remove duplicated vertices labels.
+        let mut labels: FxIndexSet<_> = labels.into_iter().map_into().collect();
+        // Sort labels.
+        labels.sort();
+
+        // Compute new graph order.
+        let order = labels.len();
+        // Initialize adjacency matrix given graph order.
+        let adjacency_matrix = DenseAdjacencyMatrix::from_elem((order, order), false);
+
+        // Assert vertex set is still sorted.
+        debug_assert!(labels.iter().is_sorted());
+        // Assert vertex set is still consistent with adjacency matrix shape.
+        debug_assert_eq!(labels.len(), adjacency_matrix.nrows());
+        // Assert adjacency matrix is still square.
+        debug_assert!(adjacency_matrix.is_square());
+
+        Self {
+            labels,
+            adjacency_matrix,
+            size: 0,
+        }
+    }
+
+    fn complete<V, I>(labels: I) -> Self
+    where
+        V: Into<String>,
+        I: IntoIterator<Item = V>,
+    {
+        // Remove duplicated vertices labels.
+        let mut labels: FxIndexSet<_> = labels.into_iter().map_into().collect();
+        // Sort labels.
+        labels.sort();
+
+        // Compute new graph order.
+        let order = labels.len();
+        // Initialize adjacency matrix given graph order.
+        let mut adjacency_matrix = DenseAdjacencyMatrix::from_elem((order, order), true);
+        // Remove self loops.
+        adjacency_matrix.diag_mut().map_inplace(|x| *x = false);
+
+        // Compute size.
+        let size = (order * (order.saturating_sub(1))) / 2;
+
+        // Assert vertex set is still sorted.
+        debug_assert!(labels.iter().is_sorted());
+        // Assert vertex set is still consistent with adjacency matrix shape.
+        debug_assert_eq!(labels.len(), adjacency_matrix.nrows());
+        // Assert adjacency matrix is still square.
+        debug_assert!(adjacency_matrix.is_square());
 
         Self {
             labels,
@@ -469,83 +543,6 @@ impl BaseGraph for UndirectedDenseAdjacencyMatrixGraph {
         debug_assert_eq!(Adj!(self, x).any(|z| z == y), f);
 
         f
-    }
-}
-
-/* Implement DefaultGraph trait. */
-impl Default for UndirectedDenseAdjacencyMatrixGraph {
-    #[inline]
-    fn default() -> Self {
-        Self {
-            labels: Default::default(),
-            adjacency_matrix: DenseAdjacencyMatrix::from_elem((0, 0), false),
-            size: 0,
-        }
-    }
-}
-
-impl DefaultGraph for UndirectedDenseAdjacencyMatrixGraph {
-    fn empty<V, I>(labels: I) -> Self
-    where
-        V: Into<String>,
-        I: IntoIterator<Item = V>,
-    {
-        // Remove duplicated vertices labels.
-        let mut labels: FxIndexSet<_> = labels.into_iter().map_into().collect();
-        // Sort labels.
-        labels.sort();
-
-        // Compute new graph order.
-        let order = labels.len();
-        // Initialize adjacency matrix given graph order.
-        let adjacency_matrix = DenseAdjacencyMatrix::from_elem((order, order), false);
-
-        // Assert vertex set is still sorted.
-        debug_assert!(labels.iter().is_sorted());
-        // Assert vertex set is still consistent with adjacency matrix shape.
-        debug_assert_eq!(labels.len(), adjacency_matrix.nrows());
-        // Assert adjacency matrix is still square.
-        debug_assert!(adjacency_matrix.is_square());
-
-        Self {
-            labels,
-            adjacency_matrix,
-            size: 0,
-        }
-    }
-
-    fn complete<V, I>(labels: I) -> Self
-    where
-        V: Into<String>,
-        I: IntoIterator<Item = V>,
-    {
-        // Remove duplicated vertices labels.
-        let mut labels: FxIndexSet<_> = labels.into_iter().map_into().collect();
-        // Sort labels.
-        labels.sort();
-
-        // Compute new graph order.
-        let order = labels.len();
-        // Initialize adjacency matrix given graph order.
-        let mut adjacency_matrix = DenseAdjacencyMatrix::from_elem((order, order), true);
-        // Remove self loops.
-        adjacency_matrix.diag_mut().map_inplace(|x| *x = false);
-
-        // Compute size.
-        let size = (order * (order.saturating_sub(1))) / 2;
-
-        // Assert vertex set is still sorted.
-        debug_assert!(labels.iter().is_sorted());
-        // Assert vertex set is still consistent with adjacency matrix shape.
-        debug_assert_eq!(labels.len(), adjacency_matrix.nrows());
-        // Assert adjacency matrix is still square.
-        debug_assert!(adjacency_matrix.is_square());
-
-        Self {
-            labels,
-            adjacency_matrix,
-            size,
-        }
     }
 }
