+++ tests/graphs/mod.rs
@@ -1,6 +1,5 @@
 mod algorithms;
 mod base;
-mod default;
 mod direction;
 mod partial_ord;
 mod path;
